# app/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API de Flask",
    "description": "Documentación de la API de Flask"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/welcome": {
      "get": {
        "summary": "Ruta de bienvenida",
        "responses": {
          "200": {
            "description": "¡Bienvenido a la API de Flask!"
          }
        }
      }
    },
    "/crear-cuenta-de-usuario": {
      "post": {
        "summary": "Crear una cuenta de usuario",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Datos del nuevo usuario",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nombre_user": {
                  "type": "string"
                },
                "email_user": {
                  "type": "string"
                },
                "password_user": {
                  "type": "string"
                },
                "sexo_user": {
                  "type": "string"
                },
                "foto_user": {
                  "type": "string"
                },
                "description_user": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario creado exitosamente"
          },
          "400": {
            "description": "Error en la creación del usuario"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Autenticación de usuario",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Credenciales de usuario",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token generado exitosamente"
          },
          "401": {
            "description": "Nombre de usuario o contraseña incorrectos"
          }
        }
      }
    },
    "/ruta-protegida": {
      "get": {
        "summary": "Ruta protegida",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario autenticado",
            "schema": {
              "type": "object",
              "properties": {
                "logged_in_as": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api": {
      "get": {
        "summary": "Llamada a una API externa",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Datos de la API externa"
          },
          "404": {
            "description": "No hay datos"
          }
        }
      }
    }
  },
  "definitions": {},
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Introduce el token JWT con el prefijo 'Bearer '"
    }
  }
}